chore(types): refactor tests for types
